+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [x.x.x] - unreleased
+### Added
+- Added support for PATCH /reports/{reportId}/definition endpoint, Update Report Definition
 
 ## [3.10.0] - 2025-12-04
 ### Added